Added global configuration options to control if users can setup scheduled checking or start the fetchmail daemon

+++ fetchmail/index.cgi
@@ -15,45 +15,49 @@ if (!&has_command($config{'fetchmail_path'})) {
 
 @conf = &parse_config_file($fetchmail_config);
 @conf = grep { $_->{'poll'} } @conf;
+if (!@conf) {
+	print "<b>$text{'index_none'}</b><p>\n";
+	}
 &show_polls(\@conf, $fetchmail_config, $remote_user_info[0]);
 
 if (@conf) {
 	# Show the fetchmail daemon form
 	print "<hr>\n";
-	print "<table width=100%>\n";
+	print &ui_buttons_start();
 
-	if (&foreign_check("cron")) {
+	if (&foreign_check("cron") && $can_cron) {
 		# Show button to manage cron job
-		print "<form action=edit_cron.cgi>\n";
-		print "<tr> <td><input type=submit ",
-		      "value='$text{'index_cron'}'></td>\n";
-		print "<td>$text{'index_crondesc'}</td> </tr></form>\n";
+		print &ui_buttons_row("edit_cron.cgi",
+				      $text{'index_cron'},
+				      $text{'index_crondesc'});
 		}
 
-	print "<tr><td>\n";
-	foreach $pf ("$remote_user_info[7]/.fetchmail.pid",
-		     "$remote_user_info[7]/.fetchmail") {
-		if (open(PID, $pf) && ($line=<PID>) &&
-		    (($pid,$interval) = split(/\s+/, $line)) && $pid &&
-		    kill(0, $pid)) {
-			$running = $pf;
-			last;
+	if ($can_daemon) {
+		print "<tr><td>\n";
+		foreach $pf ("$remote_user_info[7]/.fetchmail.pid",
+			     "$remote_user_info[7]/.fetchmail") {
+			if (open(PID, $pf) && ($line=<PID>) &&
+			    (($pid,$interval) = split(/\s+/, $line)) && $pid &&
+			    kill(0, $pid)) {
+				$running = $pf;
+				last;
+				}
+			}
+		if ($running) {
+			# daemon is running - offer to stop it
+			print &ui_buttons_row("stop.cgi",
+				$text{'index_stop'},
+				&text('index_stopmsg',$interval));
+			}
+		else {
+			# daemon isn't running - offer to start it
+			print &ui_buttons_row("start.cgi",
+				$text{'index_start'},
+				&text('index_startmsg',
+				      &ui_textbox("interval", 60, 5)));
 			}
 		}
-	if ($running) {
-		# daemon is running - offer to stop it
-		print "<form action=stop.cgi>\n";
-		print "<input type=submit value='$text{'index_stop'}'></td>\n";
-		print "<td>",&text('index_stopmsg', $interval),"</td>\n";
-		}
-	else {
-		# daemon isn't running - offer to start it
-		print "<form action=start.cgi>\n";
-		print "<input type=submit value='$text{'index_start'}'></td>\n";
-		print "<td>",&text('index_startmsg',
-			   "<input name=interval size=5 value='60'>"),"</td>\n";
-		}
-	print "</td></tr></table></form>\n";
+	print &ui_buttons_end();
 	}
 
 &ui_print_footer("/", $text{'index'});
